Switched from e.changedEvents to e.touches.
e.touches is the current list of touches. e.changedEvents is a list of the touches affected by the event. When we had a touchend event, e.changedTouches had the list of touches being ended, when those should be removed.
We keep e.changedEvents as a backup to getTouchPos in case all touches have been removed from the screen.
Also may improve finger tracking as touches persists.
Testing in Chrome, Opera and Safari on iPhone and Firefox on Android (Galaxy SII).