fixes #405: can't hit target framerate.
p5 should only paint when we it gets permission via the
requestAnimationFrame callback, which is roughly ~60 FPS.
setTimeout & setInterval are removed in this context because
they are not synchronized with when the browser finds a good
moment to draw.

If requestAnimationFrame does trigger the callback but not
enough time has passed per the user's set frameRate, drawing
is now skipped, but the browser may not be ready again until
well after the desired time, which may hamper framerate even
when ample time is available. So instead, _draw goes ahead
and renders if it's within 5ms before when it's next supposed
to render. The 5ms tolerance is arbitrary, but it seems to keep
things in check.

This approach also means that we can't paint any faster than the
browser does, which will be 60 FPS on most systems.

Also: a quick update to src/core/shim.js to rename requestDraw
to requestAnimationFrame, which is more accurate.