Run mocha specs against connect-served spec files
@iros reported issues running the p5 tests in a browser: because our
connect server was loading `./test/` as the server root, the browser
could not access the actual p5.js lib files (which were in `../lib/`,
up one level relative to `./test/`).

Changing the connect server to use `.` as its root fixes this problem,
but this issue demonstrates that our server tests and browser tests
were not executing in comparable environments. To fix this, I configured
Mocha to run the spec files via the actual spec URLs on the running
connect server, so that the command line and the browser will be
testing the exact same thing, with the same relative paths.

Changing these roots required adjusting a couple of paths in
situations where we were loading files within our tests.

Finally, I added a new grunt command: `grunt test:nobuild`, which will
lint the test files then run the mocha specs without rebuilding the
whole project: this is a developer convenience, which can be useful if
we're only making updates to the test specs themselves and p5 itself
has not changed. I also altered the existing `grunt test` command
to lint the code prior to starting the connect server, as it is only
needed for the mocha specs themselves.