Include OpenType within the built library file (fixes #781)
Concatenating OpenType.js onto the end of our module was meant only
to be a temporary measure to work around issues where the requirejs
build would blow up when trying to include OpenType. Now that we are
no longer using RequireJS, we can require the dependency directly,
bundle it into our released lib, and manage it all via NPM/browserify.

@dhowe Please review and give this a thorough field test, if you can;
this PR adds a test to validate that `loadFont` works correctly, but
I haven't done any more rigorous testing of the font code!

Issue #827 should be merged prior to merging this, or else the tests
will not run :)