Pass generated browserify bundle through derequire
**`--standalone` isn't**

Browserify's "Standalone" option is poorly named, as `--standalone`-
generated bundles cannot be consumed within any other project that is
build using browserify. The reason this is the case is that Browserify
tries to parse any "require" calls that it finds in the source.

However, once Browserify has been used to build p5, those are not
Node's `require` method anymore: they are an internal method defined
within the bundled code. Browserify trips over itself by
misinterpreting the purpose of the `require`s that it generated when
we created the bundle.

[derequire](https://github.com/calvinmetcalf/derequire) is designed
to solve this problem: it re-writes the generated bundle code to
change all instances of "require" into "_dereq_". This "hides" them
from subsequent runs of Browserify, and permits the library to be
loaded as intended.

[Browserify issue 1151](https://github.com/substack/node-browserify/issues/1151)
has more discussion of this issue, including a movement to implement
the "derequire" behavior by default when using `--standalone`, which
would (subjectively) make the flag function more predictably.

**Grunt changes**

Because `derequire` is designed either to function fully by use of its
programmatic API, or fully from the command line, we were not able to
get it integrated into the existing Grunt-Browserify task. Instead,
I have written a new, bespoke Grunt task that will

- Read the source code into Browserify (with appropriate transforms)
- Run the bundled code through `derequire` to re-write the "require"s
- Prepend the same code comment banner we have been using
- Save the processed code to a file as `lib/p5.js`

This task lives in `build/tasks/browserify.js`, using the (that is to
say, my) recommended approach of splitting more complex grunt tasks
out into individual files for easier readability and maintenance.

This task runs slightly slower than the older `grunt-browserify` task,
which I believe to be unavoidable due to the additional code
processing that is required to make the build browserifiable.

A note on Webpack: I investigated using Webpack to create the bundle,
as Webpack's UMD build target does not seem to share Browserify's
`--standalone` quirks. For now we believe that adding Webpack would
introduce more complexity and maintenance overhead than it removed,
however, so I opted for the custom Grunt task approach.

**Unresolved Issues**

This does nothing to address #785: we will likely need to adapt the
browserify task to generate separate source builds for `p5.js` and
`p5.min.js` before we can conditionally stub out behavior.

**Resolved Issues**

This fixes #818: I have copied the generated code into the
[example repository](https://github.com/geekydatamonkey/test-p5-browserify)
provided by @geekydatamonkey, and `browserify main.js -o bundle.js`
worked as expected.

**Review Requested**

Pinging @tbranyen for feedback on this implementation