fixes bug in splitTokens()
`splitTokens()` was working just as `split()`.  This fix turns the string of delimiters into a regex where each character is listed in a character class.   This is now working as expected, and I think this is a good way of doing it though I am completely open to other suggestions!