Better tests, and actually run them.
Adds the new unit test file to the test index files so that they run.

Instead of checking that particular properties are preserved by push/pop, run a deepEqual check against the render state before and after each of the operations that push is documented to preserve.  This should have a better chance of catching future subtle breaks to push/pop behavior.  Verified that the new tests actually catch the problem with _strokeSet and _fillSet that I was originally trying to fix.