Switch ascender/descender to better match processing
Processing rasterizes "d" & "p" to identify the ascender and descender
values for a font.  In p5, the ascender/descender pulled from opentype
are from the hhea table in the font header.  Those correspond to the max
ascender/descender of all glyphs.  "d" & "p" are usually smaller than
the max ascent/descent, so sTypoDescender & sTypoAscender tend to be
closer to processing's metric (for Western fonts).