Streamline Renderer2D.get()
Don't call loadPixels(); we don't need it.
When getting a single pixel value, just use getImageData with
width/height 1.

Tests pass, and there's a get() test, but it seems to never run??
I could never get it to fail, so there's something not right there.
I tested with a local sketch and things seem fine.

An attempt to help with:
https://github.com/processing/p5.js/issues/1485