Avoid .bind when assigning some methods to global object
Per #1512, there are significant slowdowns when using certain
methods in global mode, such as `random()`. Part of the origin
of this performance issue is that [methods are .bound to the
p5 instance when being assigned to the global object][1], and
the use of the .bind (which is not needed for a pure function
like `random`) impacts the method's speed. ([see explanation][2])

This PR takes a crack at a solution proposed in #1512, specifically
decorating methods which can be assigned to the global context
without the use of `.bind` with a `._context = false` property.
Core.js is then modified to respect this flag by only using `.bind`
if `method._context !== false`.

This could use some thorough testing, I've only spot-checked&mdash;
and this currently only accounts for the math utilities in the
Calculation, Noise, Random and Trigonometry files. The `_context`
argument can likely be rolled out more broadly, across any method
of `p5.prototype` which does not utilize `this`.

Unfortunately, most trigonometry functions cannot be decorated in
this way because they depend on the instance property `_angleMode`
in order to return the proper value.

[1]: https://github.com/processing/p5.js/blob/d0cf66e2d2de3d26a2d3091cb0f0471a15675249/src/core/core.js#L467
[2]: http://stackoverflow.com/questions/17638305/why-is-bind-slower-than-a-closure