Fix windowWidth and windowHeight
The `windowWidth` and `windowHeight` variables should refer to the size of the viewPort.
So we must use `window.innerWidth` and `window.innerHeight`.

Both `document.documentElement.clientWidth` and `document.body.clientHeight`
may be used as polyfill, but they refer to the dimensions of the body and the html element,
which may be greater than those of the actual viewport (!).