Fixes an issue where get() return single pixel when caller asks for image
If you ask for a section of the image like in the following code, when the video is not ready, `get()` now returns a single pixel which breaks a call to `image()`.  This could be solved with error handling in the end-user's code, but I'm adding something where `get()` returns a blank image until it's ready.  

```javascript
  var img = video.get(200,200,50,50);
  image(img, 0, 0);
```