Cache strokeStyle and fillStyle for performance
Many code paths repeatedly reset or access the stroke or the fill, for
instance when drawing shapes every call to vertex() queries the fill and
stroke styles. This seems to be quite an expensive operation, and for
almost every use-case is not changing between calls to vertex.

Here I replace all direct accesses to the canvas' drawingContext's
strokeStyle and fillStyle to go through accessors that cache the result.
This shows significant performance improvements, at least on Chrome on
Mac.