Further ensure caching when setting fill and stroke
In many cases setting the fillStyle or strokeStyle of a drawingContext
will give back a different result when querying it later. This is
because the internal representation is different, so the string
representation of the colour may end up different.

In order to ensure that setting a fill or stroke twice doesn't make a
second round-trip to the drawingContext I cache based on what we *set*
not what we get back out of the drawingContext.