fix #1678 by setting default stroke and fill colors in 3d renderer.
uses this.stroke and this.fill methods when the renderer is initialized,
so that all necessary gl context side effects also happen.

clean up gl tests:

- made the camera tests use normalMaterial(), instead of default
  stroke and fill colors.
- some tests were using basicMaterial(), which no longer exists. Changed
  to fill() or normalMaterial() as appropriate.
- remove trailing whitespaces in tests.

update conditional in _getCurShader to reflect that interleaving
immediate mode and retain mode calls requires changing shaders,
even if no change is made to fill color (so it is not sufficient
to just change the shader in fill; would require user to call
fill in between rect() and box() calls, for example).