webgl: use right lefthand coordinate system; add more camera api.
CAMERA:
can now supply a eye coord, center coord, and up vector.

COORDINATE SYSTEM:
when applying the perspective and orthographic transformations for
the camera, we now negate the y-component of the projection transform
in webgl mode. this results in a left-handed coordinate system through
and through: whereas before, the y-coordinates for positions were
inverted in the vertex shaders and in translate(). the result was
something that behaved like 2d-mode in mapping the positive y-direction
downward, but handled rotation differently. now everything should
match up and more closely match behavior from Processing, too.

so with that, changes include:
- update default vertex shader .glsl files to not negate y-coordinate
- implement custom perspective and orthographic handling of camera
  outside of the matrix class, so as to apply our little y-coordinate
  trick in the same way Processing does.
- implement a camera matrix that includes the lookAt information,
  which is persistent between frames. that matrix is then the
  first thing applied to the modelview matrix in subsequent frames;
  it will clobber any existing things in the modelview matrix if
  called by a user mid-draw.

there is more documentation and clean-up to do here, but I want to
get this in so that Stalgia can use it!