immediate mode stroking
stroke/noStroke/noFill/strokeWeight now working on immediate mode. No
noticeable performance hit on large numbers of immediate mode objects
drawn without stroke. Slight performance hit when stroking but that is
to be expected because strokes are 4 times the vertices. Even with that
performance is pretty good when drawing less than 500 four-vertex
objects.