Ignore hard to fix YUIDoc types
Tone.Signal is a type from another library so we can't "see" its type. It's only used as a return from a constructor, but constructors can't have return types in TS so we don't even emit a definition for it.
SoundObject is either p5.Sound or an object that has .connect(unit) (e.g. Amplitude, AudioNode). YUIDoc doesn't seem to have free-standing interfaces so expressing "has .connect" seems hard.