Implementing loadBytes() #2674
This commit adds the loadBytes() function to p5.js as demonstrated
during a livestream on The Coding Train. There are some remaining issues
here. For example, there are no tests. I am also not using httpDo()
which would be more of the convention here for p5 as discussed
in #2674. I am also handling errors like 404 or CORS in perhaps
an odd way and would be happy to any feedback.