Fix scaling issue when using Graphics on retina
Using `image(p5.Graphics)` on a screen with `pixelDensity()`
higher than 1 results in distorted results.

This commit fixes this issue by using the technique implemented
in `copy()`. I am however not sure if this commit breaks anything
else as  the rendering tests are sparse and I am only discovering
the codebase.

Solves: #2077