Replicating PR #188 form p5.js-website repo
Use javascript to update the document title on reference pages to
reflect the current topic being browsed. This was originally submitted
as a PR to the p5.js-website repo, but I realise that it would be
better submitted here, because these are the ultimate source of the
reference page files.