section heading
p5.js is not an application, it is a library, so this section heading was misleading. It's tempting to use the word "core" to describe this group of files but that word is already used several times in the source code and it would be confusing to use it again here with a different meaning.