update gl pipeline to set shader uniforms at render time
p5.RendererGL is updated to hold more state so that
shader uniforms can be set later, when it's time to
render a shape. then, a shader does not have to be
bound at all time to change state. this enables
things like full support for push/pop (fixes #2370).
in particular, state is added for tracking:

- whether or not the normal shader is in use
- use of lighting (boolean)
- use of specular component (boolean)

the curFillShader and curStrokeShader objects are removed
from p5.RendererGL in favor of selecting a default
shader at render time, based on current gl state
(lighting conditions, texture settings, and fill/stroke).
additionally, _userFillShader and _userStrokeShader
are introduced to indicate that a custom shader has been
set by the library user using a call to shader().
in this case, the user shader takes priority over any
default shaders. this enables the user to mix and match
calls to fill, stroke, lights, etc. to set uniforms
in their own shader.

adds p5.RendererGL._setFillUniforms and
p5.RendererGL._setStrokeUniforms to set all of the
known default uniforms for the shader. to be used
just before shape render time.

unit tests are updated to reflect this new structure;
still heavily dependent on internal state.