Add requested changes of alt tag and additional check to ensure we dont divide by zero
Implement textureMode

add IMAGE to constants.js

add textureMode example to the test folder. This shows how to use both
NORMAL and IMAGE mode using beginShape and vertex

add textureMode method. This sets a textureMode property of the renderer
to either NORMAL or IMAGE. Essentially the same function as written in
desktop processing.

update vertex to work with textureMode(IMAGE). We check for the
textureMode of the renderer and if it is image divide the uv's by the
most recently used image

add textureMode and textureImage attributes to the renderer. These help
keep track of state. TextureImage is set inside the texture() function.

unit tests for textureMode check to make sure that the renderer's mode
is correctly set

add alt tag to textureMode

add alt tag to textureMode 2nd example

add addtional check to ensure that we aren't going to divide by 0 for textureMode(IMAGE)

removed unnecessary second definition of textureMode docs