unlink touch events from mouse events in p5.Element
fixes a bug where touchEnded fires twice on tapped buttons & other
p5.Elements.

a mobile-browser quick-tap event generates a mouseup event
automatically--so having p5.Element.touchEnded generate events
for both touchended and mouseup can result in a double-firing
of the event.

I've attempted to update the documentation here, but have only
been able to test with Chrome for Android--that's why the language
in the documentation is a little non-committal.

as a precaution, I've unlinked all of the events for p5.Element.
however, for the general/global touchEnded etc. this isn't
necessary because p5 already takes precaution to make sure both
mouse/touch are not called.