Modified documentation for global `get()` function
I've added a link to `p5.Image.get()` in the documentation for
the global `get()` function, because some people (including me)
might get confused and think that you cannot use get to extract
a subimage from a `p5.Image`.

The fact that you can call `copy(image, x, y, ...)` but you can't
do the same with `get()` makes it a bit more confusing.