Use `DOMContentLoaded` instead of `unload` for initializing p5
Documentation:
https://developer.mozilla.org/en-US/docs/Web/Events/DOMContentLoaded

Quoth MDN: “A very different event load should be used only to detect a
fully-loaded page. It is an incredibly common mistake to use load where
DOMContentLoaded would be much more appropriate, so be cautious.”

DOMContentLoaded is supported in all modern browsers, and can make a
significant difference in page load performance when a page has many
external assets, such as images and iframes.