Create Step by step guide to setting up GitHub and making a first contribution.md
From the top of the proposed new file:

This is a step-by-step log of setting yourself up as a p5.js contributor, and making a first contribution, 
say a minor documentation correction on a Reference page in p5.js. This supplements the descriptions in ..

  https://github.com/processing/p5.js/blob/master/developer_docs/README.md . . . . # see section Development Process
  https://github.com/processing/p5.js/blob/master/developer_docs/preparing_a_pull_request.md
  https://github.com/processing/p5.js/blob/master/developer_docs/inline_documentation.md 
   
Environment was: MacOS Mojave 10.14.2, Chrome 70.0.3538.110, Node.js 10.15.0, p5.js 0.7.2 (more ?)

There is some detail on possible errors and warnings you might get, and the process reflects the state of the various tools as of 
Jan 2019.