Dom Drop's second callback should be called with the raw drop event
The dom library has a Element.drop() which is documented to take two
callbacks, the first should be fired for every file dropped, the second
should fire with the raw dom event. The second event was not being
called. I have fixed this bug, improved documentation, and wrote
a mocha test to cover this issue.

Resolves: #3591
See also: #3303