Fixed issue #3830
In the select method _wrapElement(res) method returns a createSelect() element without adding the element to this._elements array. The removeElements() method thus gives an error since the select element created is not a part of the _elements array. I added addElement(res, this) before the element is returned so that it's pushed to the elements array, hence fixing the issue.