updated based on discussion from p5.js contributors conference
We acknowledge that there may be times where sexually explicit or violent content may be relevant and useful for specific discussions. However, in these cases, the content should be offered respectfully, in context with explanation, and preceded by a warning so others can decide whether they'd like to view it. It is generally best to link to this material, rather than embedding directly, so others may utilize the warning if they'd like.