fix(dom): Style setting for element
This will now call the WebAPI in order to update internal state for
width height left or top rather than relying on the input provided to
the method. This way we more reliably handle when the browser silently
fails given a purely numeric value for style setting. This was dicussed
in processing/p5.js#4556.