Change distance example to be more reasonable
Previously the example converted float to integer, thus
removing all decimal places. However, the text displaying distance has exactly
one decimal place. It makes more sense to me to use that decimal place
instead of having it be zero all the time. This way the nature of
distance function is also better demonstrated (it returns a float
instead of an integer).