p5.renderer: temporarily replace the default renderer
This patch adds the p5.renderer() method to make it easier to
replace the default renderer when combining multiple sketches
into one.  When the default renderer is replaced with an off-screen
p5.Renderer object, the normal drawing calls will render into
the off screen object.  The original renderer can be restored
and the off screen renderers can then be combined into a single
image.

Signed-off-by: Trammell Hudson <hudson@trmm.net>