Allow setting the image width and height.
Allow setting the image width and height to support proper loading of SVG images as canvas images.
This allows for better scalling of image elements relative to a given screen / canvas size.

SVG files can already be loading using this method. However, currently the intrinsic size stored in an SVG file is used to determine its size, which works against the idea of using SVGs-files as scalable templates for canvas images. Allowing for parameters for the desired width and height in which the SVG should be interpreted renders the SVG file correctly and allows for easy scalability.