Ensure inline licenses are kept in p5.min.js.
For future reference, all preserved multiline comments need to have
a single line comment before them. Otherwise, in the minified file,
they will be indented incorrectly.

Bad input:
```
There is no single line comment starting with // here.
/*!
* @license MIT
* @author Votam Mann
* Cue inspired by JavaScript setTimeout, and the
* Tone.js Transport Timeline Event, 2015 tonejs.org
*/```
Output
```
/*!
    * @license MIT
    * @author Votam Mann
    * Cue inspired by JavaScript setTimeout, and the
    * Tone.js Transport Timeline Event, 2015 tonejs.org
    */
```
As it can be seen, the asteriks are indented for some unknown reason.

For some reason, designated comments are only preserved if they proceed
a function that is kept as is in the minified file. In src/dom/dom.js,
there was no way I could work out how to preserve the comment above the
internal Cue class. Because of this, I moved it underneath.

Helpful readings:
- https://github.com/mishoo/UglifyJS#output-options