change the behavior of the quad function.
The current definition of the quad function states that if there are 12 arguments, only the first 8 are used to specify the coordinates. 
I don't think this is the intended behavior, so I suggest changing it. 
It's easy to do, and it seems to be enough to remove the equals sign from the inequality sign.