Match line drawing orientation to primitive face
Line drawing in p5.js is done by creating a rectangle with both vertices offset slightly in the direction perpendicular to the line. 
In doing so, the orientation of the triangles that make it up is the opposite of the orientation of the faces of primitives (sphere, box, and so on). 
This orientation can be matched by reversing the direction of the normals used to shift the vertices in the shader.