Simplify some processing of for-loop
I thought it would be possible to simplify by nesting some of the internal processing of the for-loop that performs the interpolation processing, so I rewrote it.
I added the roles of the variables used in the loop in the form of comments. Also, I thought that "control" was more appropriate than "intermediate", so I renamed it.