Make the orientation of triangle() clockwise
The order of vertices in the triangle function is counterclockwise, unlike most other primitives.
For this reason, the direction of the normal line will be opposite to when creating a triangle by specifying the same order in immediate mode.
So I fixed it. Along with that, the specification of uv was also corrected.