Translation of the last paragraph
Finished the translation, but this is not the  final form, I'm still searching for appropriate terms for some words in the terminology, also some additional revision on the text will be done to make it perfect and make sure that there are no typos or wrong grammar.
There are a bug that occurred while writing this which is that the lists written in Arabic, have the list content in LTR while the list numbers/bullets are in RTL, I couldn't find a solution for this bug, but I'm still searching for a solution.