perform distance change logarithmically
Instead of using addition and subtraction to change distances, logarithms and multiplications are used to change distances.
This allows you to achieve the same operability regardless of scale.
In addition, we will use near and far clips for distance validation. Because the number 0.1 doesn't make sense depending on the scale.