Fixes related to touch, other small fixes
Only one touch pointer needs to be examined, so I rewrote it that way.
Also, we removed the in-canvas condition from the end condition of the touch interaction, as it is unnecessary.
I was using mag() to check whether it is positive value or not, but magSq() is ok and faster, so I rewrote it.
Lastly, the link on the Three.js site that was mentioned as a reference for orbitControl() was broken, so I rewrote it with an unbroken link. I didn't actually refer to this, but it's not good to have a broken link, so I put up a link that exists.