Minor specification change for orbitControl()
The value of mouseWheel is machine-dependent, so I'll just use only the sign to remove that effect. Also, I replaced with 'cam' where it could be replaced with 'cam'.
In addition, in the current specification, the projMatrix of the camera is not updated when scaling with ortho(), so after updating it, I rewrote it to update uPMatrix.