Overwrite previous commit README.md
I'm overwriting the previous commit, which simply fixed a typo in the README but somehow caused tests to fail. I'm putting the typo back in for now so that I can get back to the main work. I may come back later and investigate what caused the problem with the original commit.