Rename discussion.yml to 2-discussion.yml
Added a numerical prefix to the filename for the p5.js 2.0 RFC proposal form, to make it appear at the top of the template chooser, since we're prioritizing work toward p5.js 2.0 right now (and to give greater visibility to the effort, since it's new). 

Adding a numerical prefix to this filename as well, for consistency. I'm keeping the order of the original issue forms, since it's what existing contributors are already used to.