Rewrite loadTable
Implementation of p5.Table is very broken and loadTable was also very broken.
This commit does not fix the implementation fully as it still need to be
decided whether p5.Table is to be retained.

The fact that no one has noticed p5.Table and loadTable is so broken might
be a good reason to not retain them as no one is using them.